requirements.txt
accelerate==0.24.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
argilla==1.12.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.2.post2
blinker==1.6.2
bottle==0.12.25
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.4
colorama==0.4.6
commonmark==0.9.1
contourpy==1.1.0
cpm-kernels==1.0.11
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.9
Deprecated==1.2.14
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.100.0
ffmpy==0.3.1
filelock==3.9.0
filetype==1.2.0
Flask==2.3.2
fonttools==4.41.1
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.41.2
gradio_client==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
InstructorEmbedding==1.0.1
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.228
langchainplus-sdk==0.0.20
latex2mathml==3.76.0
linkify-it-py==2.0.2
loguru==0.7.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdtex2html==1.2.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.2.1
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
networkx==3.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.74
openpyxl==3.1.2
orjson==3.9.2
packaging==23.1
pandas==1.5.3
Paste @ file:///C:/ci_311/paste_1676484119636/work
pathlib==1.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.3.0
plotly==5.15.0
protobuf==4.21.12
psutil==5.9.6
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.11
pydub==0.25.1
Pygments==2.15.1
PyMySQL==1.1.0
pyodbc==5.0.1
pypandoc==1.11
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
referencing==0.30.0
regex==2023.6.3
requests==2.28.1
rfc3986==1.5.0
rich==13.0.1
rpds-py==0.9.2
rwkv==0.8.0
safetensors==0.4.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.15.0
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.35.2
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uc-micro-py==1.0.2
unstructured==0.8.0
urllib3==1.26.13
uvicorn==0.22.0
Wand==0.6.11
websockets==11.0.3
Werkzeug==2.3.6
win32-setctime==1.1.0
wrapt==1.14.1
wudao==1.3.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
